+++ pkgs/firehose/lib/src/github.dart
@@ -27,6 +27,10 @@ class Github {
   /// The PR (or issue) number.
   String? get issueNumber => _env['ISSUE_NUMBER'];
 
+  /// Any labels applied to this PR.
+  List<String> get prLabels =>
+      _env.containsKey('PR_LABELS') ? _env['PR_LABELS']!.split(',') : [];
+
   /// The commit SHA that triggered the workflow.
   String? get sha => _env['GITHUB_SHA'];